@@ -1,1 +1,1 @@
-package com.blog.cron; import com.blog.service.MsmService;import lombok.extern.slf4j.Slf4j;import org.springframework.scheduling.annotation.Scheduled;import org.springframework.stereotype.Component;import javax.annotation.Resource;/** * @Description * @ClassName Task1 * @Author User * @date 2020.06.07 12:24 */@Component@Slf4jpublic class Task {    @Resource    private MsmService msmService;    @Scheduled(cron = "0 0 15,16,17,18 * * ?")    public void sayWord() {        msmService.sendEmail("2249415510@qq.com","记得收拾狗窝；记得烧水；记得洗衣服、晒衣服");        log.info("success send message!");    }    @Scheduled(cron = "0 30-59/3 11 * * ?")    public void scheduledTask() {      msmService.sendEmail("2249415510@qq.com","老子要就寝了，快睡觉！快睡觉！快睡觉！");    }}
+package com.blog.cron; import com.blog.service.MsmService;import jakarta.annotation.Resource;import lombok.extern.slf4j.Slf4j;import org.springframework.scheduling.annotation.Scheduled;import org.springframework.stereotype.Component;/** * @Description * @ClassName Task1 * @Author User * @date 2020.06.07 12:24 */@Component@Slf4jpublic class Task {    @Resource    private MsmService msmService;    @Scheduled(cron = "0 0 15,16,17,18 * * ?")    public void sayWord() {        msmService.sendEmail("2249415510@qq.com","记得收拾狗窝；记得烧水；记得洗衣服、晒衣服");        log.info("success send message!");    }    @Scheduled(cron = "0 30-59/3 11 * * ?")    public void scheduledTask() {      msmService.sendEmail("2249415510@qq.com","老子要就寝了，快睡觉！快睡觉！快睡觉！");    }}
