@@ -1,23 +1,16 @@
 package com.blog.service.impl;
 
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.blog.domain.entity.Label;
 import com.blog.domain.entity.LabelArticle;
-import com.blog.domain.entity.Type;
 import com.blog.mapper.LabelArticleMapper;
 import com.blog.mapper.LabelMapper;
-import com.blog.mapper.TypeMapper;
 import com.blog.service.ILabelArticleService;
-import com.blog.service.ILabelService;
-import com.blog.service.ITypeService;
+import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
-
-import javax.annotation.Resource;
 import java.util.List;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 @Service
 public class LabelArticleServiceImpl  extends ServiceImpl<LabelArticleMapper, LabelArticle> implements ILabelArticleService {
